[utils/channel] Split channel.go into reader_writer.go and ready.go
  Separate the two unrelated concepts — context-aware channel
  read/write wrappers and the Ready signaling primitive — into
  their own files for easier navigation and review.

  No logic changes; pure file reorganization.
