+++ utils/channel/reader_writer.go
@@ -8,8 +8,6 @@ package channel
 
 import (
 	"context"
-	"sync"
-	"sync/atomic"
 	"time"
 )
 
@@ -43,20 +41,6 @@ type (
 	WithContext[T any] interface {
 		Context() context.Context
 		WithContext(ctx context.Context) ReaderWriter[T]
-	}
-
-	// Ready supports waiting for readiness and notifying of readiness.
-	// It also supports closing to release waiters.
-	// Ready uses atomic pointer indirection to allow safe Reset() while other
-	// goroutines may be waiting on the channels. This prevents race conditions
-	// when the internal state is replaced.
-	Ready struct {
-		ready atomic.Pointer[ready]
-	}
-	ready struct {
-		ready  chan any
-		closed chan any
-		once   sync.Once
 	}
 
 	//nolint:containedctx // holding the context is required.
@@ -153,67 +137,4 @@ func (c *channel[T]) Write(value T) bool {
 	case c.output <- value:
 		return true
 	}
-}
-
-// NewReady instantiate a new Ready.
-func NewReady() *Ready {
-	r := &Ready{}
-	r.ready.Store(newReady())
-	return r
-}
-
-func newReady() *ready {
-	return &ready{
-		ready:  make(chan any),
-		closed: make(chan any),
-	}
-}
-
-// Reset resets the object to be reused.
-func (r *Ready) Reset() {
-	rr := r.ready.Load()
-	rr.once.Do(func() {
-		close(rr.closed)
-	})
-	// We only set a new internal ready to replace the one we closed.
-	// If another process already replaced it, we can continue.
-	r.ready.CompareAndSwap(rr, newReady())
-}
-
-// SignalReady signals readiness.
-func (r *Ready) SignalReady() {
-	rr := r.ready.Load()
-	rr.once.Do(func() {
-		close(rr.ready)
-	})
-}
-
-// Close notifies of closing.
-func (r *Ready) Close() {
-	rr := r.ready.Load()
-	rr.once.Do(func() {
-		close(rr.closed)
-	})
-}
-
-// WaitForReady returns true if the object is ready,
-// or false if it is closed or the context ended before that.
-func (r *Ready) WaitForReady(ctx context.Context) bool {
-	return WaitForAllReady(ctx, r)
-}
-
-// WaitForAllReady returns true if all objects are ready,
-// or false if one of them is closed or the context ended before that.
-func WaitForAllReady(ctx context.Context, ready ...*Ready) bool {
-	for _, r := range ready {
-		rr := r.ready.Load()
-		select {
-		case <-ctx.Done():
-			return false
-		case <-rr.closed:
-			return false
-		case <-rr.ready:
-		}
-	}
-	return true
 }
